# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RSS feed management app",
        "title": "Gopherss API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/entry": {
            "get": {
                "description": "get entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "List entries",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "feed",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "read",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ascending",
                            "descending"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "AscendingSort",
                            "DescendingSort"
                        ],
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "starred",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Entry"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "update entries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Update entries",
                "parameters": [
                    {
                        "description": "Update entries",
                        "name": "entries",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEntriesInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/entry/mark": {
            "post": {
                "description": "mark entries as read/unread up to a timestamp",
                "tags": [
                    "entry"
                ],
                "summary": "Mark as read/unread",
                "parameters": [
                    {
                        "enum": [
                            "read",
                            "unread"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ReadEntryStatus",
                            "UnreadEntryStatus"
                        ],
                        "name": "as",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "feed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/entry/{id}": {
            "get": {
                "description": "get entry",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Get entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Entry"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "update entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Update entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update entry",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEntryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Entry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/entry/{id}/mark": {
            "post": {
                "description": "mark entry as read/unread",
                "tags": [
                    "entry"
                ],
                "summary": "Mark as read/unread",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "read",
                            "unread"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "ReadEntryStatus",
                            "UnreadEntryStatus"
                        ],
                        "name": "as",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Entry": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "collectedOn": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "feedId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isMuted": {
                    "type": "boolean"
                },
                "isRead": {
                    "type": "boolean"
                },
                "isStarred": {
                    "type": "boolean"
                },
                "link": {
                    "$ref": "#/definitions/url.URL"
                },
                "originalId": {
                    "type": "string"
                },
                "publishedOn": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateEntriesInput": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "isMuted": {
                    "type": "boolean"
                },
                "isRead": {
                    "type": "boolean"
                },
                "isStarred": {
                    "type": "boolean"
                }
            }
        },
        "model.UpdateEntryInput": {
            "type": "object",
            "properties": {
                "isMuted": {
                    "type": "boolean"
                },
                "isRead": {
                    "type": "boolean"
                },
                "isStarred": {
                    "type": "boolean"
                }
            }
        },
...
        "url.URL": {
            "type": "object",
            "properties": {
                "forceQuery": {
                    "description": "append a query ('?') even if RawQuery is empty",
                    "type": "boolean"
                },
                "fragment": {
                    "description": "fragment for references, without '#'",
                    "type": "string"
                },
                "host": {
                    "description": "host or host:port (see Hostname and Port methods)",
                    "type": "string"
                },
                "omitHost": {
                    "description": "do not emit empty host (authority)",
                    "type": "boolean"
                },
                "opaque": {
                    "description": "encoded opaque data",
                    "type": "string"
                },
                "path": {
                    "description": "path (relative paths may omit leading slash)",
                    "type": "string"
                },
                "rawFragment": {
                    "description": "encoded fragment hint (see EscapedFragment method)",
                    "type": "string"
                },
                "rawPath": {
                    "description": "encoded path hint (see EscapedPath method)",
                    "type": "string"
                },
                "rawQuery": {
                    "description": "encoded query values, without '?'",
                    "type": "string"
                },
                "scheme": {
                    "type": "string"
                },
                "user": {
                    "description": "username and password information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/url.Userinfo"
                        }
                    ]
                }
            }
        },
        "url.Userinfo": {
            "type": "object"
        }
    }
}